Initial Commit
This initial commit contains a lot of code, and the majority of it was
simply pulled out of the https://github.com/sallamander/dl-playground
repository in order to build a stand-alone package that included the
Keras-centric training / evaluation code for PyTorch networks.

Major changes / additions added in the initial commit largely reolve around
the addition of the `backend.py` in order to allow for Keras to be used
with a `numpy` backend, rather than requiring tensorflow or some other
backend that would add an additional dependency, or creating a PyTorch
backend, which would be a much larger effort.