Add initial, limited implementation of multi-output training
This commit allows for multi-output training in (very) limited way (at least
compared to what is possible with the Keras Model class). In particular, users
must specify multiple outputs via the `__init__` of the Model class using the
`n_outputs` parameter.  Then, the targets passed to the network must be a list
of torch.Tensor objects whose length matches `n_outputs` passed to the init.
The network passed to the `Model.__init__`, must return `n_outputs` number of
outputs. The objective function will be applied to each pair of (targets,
outputs) and summed with equal weights, and each metric will be applied to each
pair of (targets, outputs). Subsequent commits will make updates and make
extensions to this limited functionality to bring it more in line with what the
Keras Model class offers.